refactor: refactor file name and structure
