+++ src/app/admin/ensiklopedia/page.tsx
@@ -1,5 +1,5 @@
-import { AdminEnsiklopediaPage } from '@/components/admin/ensiklopedia/admin-ensiklopedia-page';
+import { AdminEnsiklopediaContent } from '@/components/admin/ensiklopedia/admin-ensiklopedia-content';
 
 export default function AdminEnsiklopediaRoute() {
-  return <AdminEnsiklopediaPage />;
+  return <AdminEnsiklopediaContent />;
 }
